# caff-backend/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Kéne egy csapatnév Számítógép-biztonság HF",
        "version": "1.0.0",
        "title": "KeneEgyCsapatnev"
    },
    "host": "localhost:8080",
    "tags": [
        {
            "name": "auth",
            "description": "Authorization-related"
        },
        {
            "name": "user",
            "description": "user-related"
        },
        {
            "name": "caff",
            "description": "caff-related"
        },
        {
            "name": "admin",
            "description": "admin-related"
        }
    ],
    "schemes": [
        "http"
    ],
    "paths": {
        "/auth/signup": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Signs up a new user",
                "operationId": "registerUser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "The user to sign up",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SignupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User registrated successfully!",
                        "schema": {
                            "$ref": "#/definitions/MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Error: Username is already taken! or Email is already in use!",
                        "schema": {
                            "$ref": "#/definitions/MessageResponse"
                        }
                    }
                }
            }
        },
        "/auth/signin": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Signs in an existing user",
                "operationId": "authenticateUser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "The user to log in",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/LoginResponse"
                        }
                    }
                }
            }
        },
        "/user/{username}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Show the user's profile",
                "operationId": "getUserProfile",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "The user's username",
                        "required": true,
                        "type": "string"
                    },
					{
						"name": "Authorization",
						"in": "header",
						"description": "Bearer {jwt-token}",
						"type": "string"
					}],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
					"403": {
						"description": "Other users are not available",
                        "schema": {
                            "$ref": "#/definitions/MessageResponse"
                        }
					}
                }
            },
            "put": {
                "tags": [
                    "user"
                ],
                "summary": "Edit the user's profile",
                "operationId": "updateUserProfile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "The user's username",
                        "required": true,
                        "type": "string"
                    },
					{
						"name": "Authorization",
						"in": "header",
						"description": "Bearer {jwt-token}",
						"type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "The updated values for the user",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User was updated successfully",
                        "schema": {
                            "$ref": "#/definitions/MessageResponse"
                        }
                    },
					"403": {
						"description": "Other users are not available",
                        "schema": {
                            "$ref": "#/definitions/MessageResponse"
                        }
					}
                }
            }
        },
        "/caff/{caffid}/comment": {
            "put": {
                "tags": [
                    "caff"
                ],
                "summary": "Add a comment to the CAFF file",
                "operationId": "commentCAFF",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "caffid",
                        "in": "path",
                        "description": "The commented CAFF file's id",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
					{
						"name": "Authorization",
						"in": "header",
						"description": "Bearer {jwt-token}",
						"type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "The comment itself",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Comment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comment added successfully",
                        "schema": {
                            "$ref": "#/definitions/MessageResponse"
                        }
                    },
                    "400": {
                        "description": "CAFF not found",
                        "schema": {
                            "$ref": "#/definitions/MessageResponse"
                        }
                    }
                }
            }
        },
        "/caff/upload": {
            "post": {
                "tags": [
                    "caff"
                ],
                "summary": "Upload a CAFF file",
                "operationId": "uploadCAFF",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
					{
						"name": "Authorization",
						"in": "header",
						"description": "Bearer {jwt-token}",
						"type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "The CAFF file",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CAFFUpLoad"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CAFF uploaded successfully",
                        "schema": {
                            "$ref": "#/definitions/MessageResponse"
                        }
                    }
                }
            }
        },
        "/caff/{caffid}/download": {
            "get": {
                "tags": [
                    "caff"
                ],
                "summary": "Download a CAFF file",
                "operationId": "downloadCAFF",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "caffid",
                        "in": "path",
                        "description": "The to-be-downloaded CAFF file's id",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
					{
						"name": "Authorization",
						"in": "header",
						"description": "Bearer {jwt-token}",
						"type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/CAFFDownLoad"
                        }
                    },
                    "400": {
                        "description": "CAFF not found",
                        "schema": {
                            "$ref": "#/definitions/MessageResponse"
                        }
                    }
                }
            }
        },
        "/caff/{caffid}": {
            "get": {
                "tags": [
                    "caff"
                ],
                "summary": "View a CAFF file",
                "operationId": "viewCAFF",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "caffid",
                        "in": "path",
                        "description": "The viewn CAFF file's id",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
					{
						"name": "Authorization",
						"in": "header",
						"description": "Bearer {jwt-token}",
						"type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/CAFFPreview"
                        }
                    },
                    "400": {
                        "description": "CAFF not found",
                        "schema": {
                            "$ref": "#/definitions/MessageResponse"
                        }
                    }
                }
            }
        },
        "/caff": {
            "get": {
                "tags": [
                    "caff"
                ],
                "summary": "Show all CAFF files",
                "operationId": "listCAFF",
                "produces": [
                    "application/json"
                ],
                "parameters": [
					{
						"name": "Authorization",
						"in": "header",
						"description": "Bearer {jwt-token}",
						"type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/CAFFPreview"
                            }
                        }
                    }
                }
            }
        },
        "/caff/find/{namefilter}": {
            "get": {
                "tags": [
                    "caff"
                ],
                "summary": "Search for a CAFF by name",
                "operationId": "findCAFF",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "namefilter",
                        "in": "path",
                        "description": "A part of the searched CAFF's name",
                        "required": true,
                        "type": "string"
                    },
					{
						"name": "Authorization",
						"in": "header",
						"description": "Bearer {jwt-token}",
						"type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/CAFFPreview"
                            }
                        }
                    },
                    "400": {
                        "description": "CAFFs not found",
                        "schema": {
                            "$ref": "#/definitions/MessageResponse"
                        }
                    }
                }
            }
        },
        "/admin/user/{userid}": {
            "get": {
                "tags": [
                    "admin"
                ],
                "summary": "Show the selected user's profile",
                "operationId": "getAdminUserProfile",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "userid",
                        "in": "path",
                        "description": "The selected user's id",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
					{
						"name": "Authorization",
						"in": "header",
						"description": "Bearer {jwt-token}",
						"type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/MessageResponse"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "admin"
                ],
                "summary": "Edit the selected user's profile",
                "operationId": "updateAdminUserProfile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "userid",
                        "in": "path",
                        "description": "The selected user's id",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
					{
						"name": "Authorization",
						"in": "header",
						"description": "Bearer {jwt-token}",
						"type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "The updated values for the user",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User was updated successfully",
                        "schema": {
                            "$ref": "#/definitions/MessageResponse"
                        }
                    },
                    "400": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/MessageResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "admin"
                ],
                "summary": "Delete the selected user's profile",
                "operationId": "deleteAdminUserProfile",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "userid",
                        "in": "path",
                        "description": "The selected user's id",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
					{
						"name": "Authorization",
						"in": "header",
						"description": "Bearer {jwt-token}",
						"type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/MessageResponse"
                        }
                    },
                    "400": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/MessageResponse"
                        }
                    }
                }
            }
        },
        "/admin/caff/{caffid}": {
            "put": {
                "tags": [
                    "admin"
                ],
                "summary": "Edit the selected CAFF file",
                "operationId": "updateAdminCAFF",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "caffid",
                        "in": "path",
                        "description": "The selected CAFF file's id",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
					{
						"name": "Authorization",
						"in": "header",
						"description": "Bearer {jwt-token}",
						"type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "The updated values for the CAFF preview",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CAFFPreview"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CAFF preview was updated successfully",
                        "schema": {
                            "$ref": "#/definitions/MessageResponse"
                        }
                    },
                    "400": {
                        "description": "CAFF file not found",
                        "schema": {
                            "$ref": "#/definitions/MessageResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "admin"
                ],
                "summary": "Delete the selected CAFF file",
                "operationId": "deleteAdminCAFF",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "caffid",
                        "in": "path",
                        "description": "The selected CAFF file's id",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
					{
						"name": "Authorization",
						"in": "header",
						"description": "Bearer {jwt-token}",
						"type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CAFF file deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/MessageResponse"
                        }
                    },
                    "400": {
                        "description": "CAFF file not found",
                        "schema": {
                            "$ref": "#/definitions/MessageResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "CAFFDownLoad": {
            "type": "object",
            "required": [
                "name",
                "uri",
				"type"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "description": "The name of the CAFF file"
                },
				"uri": {
					"type": "string",
					"description": "The path of the CAFF file"
				},
                "type": {
                    "type": "string",
                    "description": "The CAFF file's type"
                }
            }
        },
		"CAFFUpLoad": {
            "type": "object",
            "required": [
                "name",
                "file"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "description": "The name of the CAFF file"
                },
                "file": {
                    "type": "object",
                    "description": "The CAFF file"
                }
            }
        },
        "CAFFPreview": {
            "type": "object",
            "required": [
                "id",
                "name",
                "imguri"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "name": {
                    "type": "string",
                    "description": "The name of the CAFF file"
                },
                "uri": {
                    "type": "string",
                    "description": "The path of the image generated from first CIFF in CAFF"
                },
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Comment"
                    },
                    "description": "The comments on the CAFF preview"
                }
            }
        },
        "Comment": {
            "type": "object",
            "required": [
                "username",
                "comment"
            ],
            "properties": {
                "username": {
                    "type": "string",
                    "description": "The commenting user's username"
                },
                "comment": {
                    "type": "string",
                    "description": "The comment itself"
                }
            }
        },
        "User": {
            "type": "object",
            "required": [
                "user_id",
                "username",
                "email",
                "password",
				"roles"
            ],
            "properties": {
                "user_id": {
                    "type": "integer",
                    "format": "int64"
                },
                "username": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "LoginRequest": {
            "type": "object",
            "required": [
                "username",
                "password"
            ],
            "properties": {
                "username": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "LoginResponse": {
            "type": "object",
            "required": [
                "jwt",
                "type",
                "id",
                "username",
                "email",
                "roles"
            ],
            "properties": {
                "jwt": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "username": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "MessageResponse": {
            "type": "object",
            "required": [
                "message"
            ],
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "SignupRequest": {
            "type": "object",
            "required": [
                "username",
                "email",
                "password"
            ],
            "properties": {
                "username": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    }
}